Changed type of value on User ID

+++ src/main/java/com/jp/prestamosjava/dao/DebtDao.java
@@ -16,7 +16,7 @@ public class DebtDao {
     public void save(DebtEntity debt){
         this.entityManager.persist(debt);
     }
-    public List<DebtEntity> debtsList(int lender){
+    public List<DebtEntity> debtsList(Long lender){
         String jpql = "SELECT d FROM DebtEntity d WHERE d.lender = :lender";
         TypedQuery<DebtEntity> query = entityManager.createQuery(jpql, DebtEntity.class);
         query.setParameter("lender", lender);
